feat: Create ProjectEmployeeService with addEntityToAssociation method
This commit introduces a new service called ProjectEmployeeService, which contains a single method called addEntityToAssociation. This method is intended to be used by both the EmployeeController and ProjectController to add associations between projects and employees. Additionally, a new method called getAvailableProjects was added to the EmployeeServiceImpl for retrieving all available projects for a given employee, and a new method called getAvailableEmployee was added to the ProjectServiceImpl for retrieving all available employees for a given project.